+++ src/main/java/org/sopt/domain/article/controller/ArticleController.java
@@ -55,25 +55,6 @@ public class ArticleController {
         CommentResponseDto response = commentService.createComment(articleId, commentCreateRequestDto);
         return BaseResponse.ok(CommentSuccessCode.CREATE_COMMENT_SUCCESS.getMsg(), response);
     }
-
-    @PutMapping("/comments/{commentId}")
-    public BaseResponse<CommentResponseDto> updateComment(@PathVariable Long commentId,
-                                                          @Valid @RequestBody CommentUpdateRequestDto commentUpdateRequestDto) {
-        CommentResponseDto response = commentService.updateComment(commentId, commentUpdateRequestDto);
-        return BaseResponse.ok(CommentSuccessCode.UPDATE_COMMENT_SUCCESS.getMsg(), response);
-    }
-
-    @DeleteMapping("/comments/{commentId}")
-    public BaseResponse<Void> deleteComment(@PathVariable Long commentId) {
-        commentService.deleteComment(commentId);
-        return BaseResponse.ok(CommentSuccessCode.DELETE_COMMENT_SUCCESS.getMsg(), null);
-    }
-
-    @GetMapping("/comments/{commentId}")
-    public BaseResponse<CommentResponseDto> getComment(@PathVariable Long commentId){
-        CommentResponseDto response = commentService.getComment(commentId);
-        return BaseResponse.ok(CommentSuccessCode.GET_COMMENT_SUCCESS.getMsg(), response);
-    }
 }
